implement useCaptureThumbnail hook
